feat(android): add relay and gatt proxy settings

+++ app/android/TelinkBleMesh/TelinkBleMesh/TelinkBleMeshLib/src/main/java/com/telink/ble/mesh/core/message/config/NodeIdentity.java
@@ -1,14 +1,14 @@
 /********************************************************************************************************
- * @file     NodeIdentity.java 
+ * @file NodeIdentity.java
  *
- * @brief    for TLSR chips
+ * @brief for TLSR chips
  *
- * @author	 telink
- * @date     Sep. 30, 2010
+ * @author telink
+ * @date Sep. 30, 2010
  *
- * @par      Copyright (c) 2010, Telink Semiconductor (Shanghai) Co., Ltd.
+ * @par Copyright (c) 2010, Telink Semiconductor (Shanghai) Co., Ltd.
  *           All rights reserved.
- *           
+ *
  *			 The information contained herein is confidential and proprietary property of Telink 
  * 		     Semiconductor (Shanghai) Co., Ltd. and is available under the terms 
  *			 of Commercial License Agreement between Telink Semiconductor (Shanghai) 
@@ -17,23 +17,23 @@
  *
  * 			 Licensees are granted free, non-transferable use of the information in this 
  *			 file under Mutual Non-Disclosure Agreement. NO WARRENTY of ANY KIND is provided. 
- *           
+ *
  *******************************************************************************************************/
 package com.telink.ble.mesh.core.message.config;
 
 public enum NodeIdentity {
-    STOPPED(0, "Node Identity for a subnet is stopped"),
+    STOPPED((byte) 0, "Node Identity for a subnet is stopped"),
 
-    RUNNING(1, "Node Identity for a subnet is running"),
+    RUNNING((byte) 1, "Node Identity for a subnet is running"),
 
-    UNSUPPORTED(2, "Node Identity is not supported"),
+    UNSUPPORTED((byte) 2, "Node Identity is not supported"),
 
-    UNKNOWN_ERROR(0xFF, "unknown error");
+    UNKNOWN_ERROR((byte) 0xFF, "unknown error");
 
-    public final int code;
+    public final byte code;
     public final String desc;
 
-    NodeIdentity(int code, String desc) {
+    NodeIdentity(byte code, String desc) {
         this.code = code;
         this.desc = desc;
     }
